blog_tyk_nu/requirements.txt
Django==1.8.2
django-bootstrap3==5.4.0
django-taggit==0.17.1
django-taggit-templatetags==0.2.5
django-haystack==2.4.0
Whoosh==2.7.0
